ExtendedDropdonwMenu for employee and type + FullEstate model creation + layouts

+++ app/src/main/java/com/openclassrooms/realestatemanager/database/model/Type.kt
@@ -9,7 +9,7 @@ data class Type(
 
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "type_id")
-    val typeId: Long,
+    val typeId: Int,
 
     @ColumnInfo(name = "name")
     var typeName: String,
